backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.30",
    "uvicorn>=0.40.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "psycopg2>=2.9.11",
]
